+++ serve.js
@@ -86,6 +86,14 @@ async function setAccessToken() {
 ///////////////////// MAIN APP ENTRY POINT /////////////////////
 
 /**
+ * Register invalid Github POST requests to handler via /github endpoint.
+ * Failed spoof attempts may end up here but most likely it will be unsupported webhook events.
+ */
+handler.on('error', function (err) {
+  console.log('Error:', err.message);
+});
+
+/**
  * Callback to deal with POST requests from /github endpoint, authenticates as app and passes on
  * request to handler.
  * @param {Object} req - Request object.
@@ -97,17 +105,13 @@ srv.post('/github', async (req, res, next) => {
    console.log('Post received')
    let id = req.header('x-github-hook-installation-target-id');
    if (id != process.env.GITHUB_APP_IDENTIFIER) { next(); return; }  // Not for us; move on
-   await setAccessToken();
-   log.extend('event')('X-GitHub-Event: %s', req.header('X-GitHub-Event'));
-   handler(req, res, () => res.end('ok'));
-});
-
-/**
- * Register invalid Github POST requests to handler via /github endpoint.
- * Failed spoof attempts may end up here but most likely it will be unsupported webhook events.
- */
-handler.on('error', function (err) {
-  console.log('Error:', err.message);
+   if (req.header('X-GitHub-Event') in supportedEvents) {
+      await setAccessToken();
+      handler(req, res, () => res.end('ok'));
+   } else {
+      log('GitHub Event "%s" not supported', req.header('X-GitHub-Event'));
+      res.sendStatus(400);
+   }
 });
 
 
@@ -222,7 +226,7 @@ function runTests(job) {
 
    // Go ahead with tests
    const sha = job.data['sha'];
-   const repoPath = getRepoPath(job.data.repo);
+   const repoPath = lib.getRepoPath(job.data.repo);
    const logName = path.join(config.dataPath, 'reports', sha, `std_output-${lib.shortID(sha)}.log`);
    let fcn = lib.fullpath(config.test_function);
    debug('starting test child process %s', fcn);
@@ -278,7 +282,7 @@ function runTests(job) {
 
 function prepareEnv(job, callback) {
    log('Preparing environment for job #%g', job.id)
-   const repoPath = getRepoPath(job.data.repo);
+   const repoPath = lib.getRepoPath(job.data.repo);
    switch (config.setup_function) {
       case undefined:
          // run some basic git commands
@@ -339,41 +343,6 @@ function checkout(repoPath, ref) {
       verify(shell.exec('git status'));
    }
    shell.popd();
-}
-
-
-/**
- * Lists the submodules within a Git repository.  If none are found null is returned.
- * @param {String} repoPath - The path of the repository
- * @returns {Array} A list of submodule names, or null if none were found
- */
-function listSubmodules(repoPath) {
-   if (!shell.which('git')) { throw new Error('Git not found on path'); }
-   shell.pushd(repoPath);
-   let listModules = 'git config --file .gitmodules --get-regexp path | awk \'{ print $2 }\'';
-   const modules = shell.exec(listModules);
-   shell.popd();
-   return (!modules.code && modules.stdout !== '')? modules.split('\n') : null;
-}
-
-/**
- * Get the corresponding repository path for a given repo.  The function first checks the settings.
- * If the `repos` field doesn't exist, the path in ENV is used.  If the name is not a key in the
- * `repos` object then we check each repo path for submodules and return the first matching
- * submodule path.  Otherwise returns null.
- * @param {String} name - The name of the repository
- * @returns {String} The repository path if found
- */
-function getRepoPath(name) {
-   if (!config.repos) { return process.env['REPO_PATH']; }  // Legacy, to remove
-   if (config.repos.name) { return config.repos.name; }  // Found path, return
-   for (let repo of config.repos) {
-      let modules = listSubmodules(repo);
-      if (modules && modules.includes(name)) {
-         // If the repo is a submodule, modify path
-         return repo + path.sep + name;
-      }
-   }
 }